Figdb: make State threadsafe
Set and Archive are implicitly thread safe because they read/write from the cache/store all at once, and the cache and keystore are thread safe. State read/writes actually involve multiple read/writes from cache/store as the path is walked; therefore, we make the externally callable methods of State threadsafe via a mutex. This allows many threads to read from state so long as nothing is writing to it, but only one thread at a time may write to state.